Rename controller files
컨트롤러 로직 파일들의 이름 변경

- user.ctrl.ts -> user.controller.ts
- studyUser.ctrl.ts -> user.controller.ts
